feat(emit-racket-oo): tighten integer contracts via enum-typedef signedness
Investigation (item3-gapb-findings.md) refuted the design's premise: no
integer primitive falls through to any/c — the ObjC extractor canonicalises
NSInteger->int64. The real gap was map_contract's Alias arm, which mapped
every framework-prefixed enum typedef to exact-nonnegative-integer?, wrongly
rejecting negative cases in 716 signed enum typedefs (NSComparisonResult
.orderedAscending = -1, etc.).

- map_contract's Alias arm now consults the alias's underlying_primitive:
  a signed int* width emits exact-integer?; unsigned/absent keeps
  exact-nonnegative-integer?.
- Defence-in-depth: map_contract's primitive arm and the FFI mapper's
  racket_ffi_type_for_primitive both gain nsinteger/nsuinteger arms, kept
  in lockstep so the contract and FFI type mappers recognise the same set.
- test_fixtures.rs type_int() fixed from the raw name "NSInteger" (which the
  real extractor never emits) to canonical "int64" — removes a fixture that
  misrepresented extractor output.
- Golden snapshots regenerated: 716-typedef family flips
  exact-nonnegative-integer? -> exact-integer? where signed; TestKit `tag`
  contracts gain exact-integer? and _int64 (previously any/c and _pointer).

+++ generation/crates/emit/src/test_fixtures.rs
@@ -116,7 +116,7 @@ fn type_int() -> TypeRef {
     TypeRef {
         nullable: false,
         kind: TypeRefKind::Primitive {
-            name: "NSInteger".to_string(),
+            name: "int64".to_string(),
         },
     }
 }
